requirements.txt
streamlit
openai
pdfminer.six
requests

